requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyYAML==6.0.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
Werkzeug==3.1.3
